2413: Improve compensation. (#979)
* 2413: Improve compensation.

* 2413: old data expand 100 times

+++ src/main/java/shamu/company/employee/dto/CompensationDto.java
@@ -1,6 +1,5 @@
 package shamu.company.employee.dto;
 
-import java.math.BigInteger;
 import lombok.Data;
 
 @Data
@@ -8,7 +7,7 @@ public class CompensationDto {
 
   private String id;
 
-  private BigInteger wage;
+  private Double wage;
 
   private CompensationOvertimeStatusDto overtimeStatus;
